requirements.txt
aiogram>=3.15.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
typing_extensions>=4.6.0
